requirements.txt
lxml==4.4.2
Pillow==7.0.0
regex==2020.1.8
